utils/config_utils.py
import yaml

def load_config(file_path: str) -> dict:
    """
    Load the YAML configuration file.
    
    Args:
        file_path (str): Path to the YAML file.
        
    Returns:
        dict: Parsed configuration as a dictionary.
    """
    with open(file_path, 'r') as file:
        return yaml.safe_load(file)

def get_config(config: dict, *keys):
    """
    Retrieve a specific configuration value using keys.
    
    Args:
        config (dict): Loaded configuration dictionary.
        *keys: Sequence of keys to traverse the configuration.
        
    Returns:
        The value associated with the given keys, or None if not found.
    """
    value = config
    for key in keys:
        if isinstance(value, dict) and key in value:
            value = value[key]
        else:
            return None
    return value