requirements.txt
langchain
langchain-openai
langchain-community
langchain-anthropic
langchain-cohere
openai
datasets
langchain-google-genai
google-generativeai
beautifulsoup4
pyyaml